+++ R/stat-panel-counts.R
@@ -1,8 +1,10 @@
 #' Number of observations in a plot panel
 #'
 #' \code{stat_panel_counts()} counts the number of observations in each panel.
-#' By default it adds a text label to the top right corner of each panel.
-#' Grouping is ignored.
+#' \code{stat_group_counts()} counts the number of observations in each group.
+#' By default they add one or more text labels to the top right corner of each
+#' panel. Grouping is ignored by \code{stat_panel_counts()}. If no grouping
+#' exists, the two statistics behave similarly.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
 #'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be
@@ -26,9 +28,9 @@
 #' @param label.x,label.y \code{numeric} Coordinates (in npc units) to be used
 #'   for absolute positioning of the labels.
 #'
-#' @details This statistic can be used to automatically count observations in
-#'   each panel of a plot, and by default add these counts as text labels. This
-#'   statistic, unlike \code{stat_quadrant_counts()} requires only one of
+#' @details These statistics can be used to automatically count observations in
+#'   each panel of a plot, and by default add these counts as text labels. These
+#'   statistics, unlike \code{stat_quadrant_counts()} requires only one of
 #'   \emph{x} or \emph{y} aesthetics and can be used together with statistics
 #'   that have the same requirement, like \code{stat_density()}.
 #'
@@ -36,17 +38,17 @@
 #'   facets even with free limits for \emph{x} and \emph{y} axes, the location
 #'   of the labels is consistent across panels. This is achieved by use of
 #'   \code{geom = "text_npc"} or \code{geom = "label_npc"}. To pass the
-#'   positions in native data units, pass \code{geom = "text"} explicitly as
-#'   argument. A vector with the same length as the number of panels in the
-#'   figure can be used if needed.
+#'   positions in native data units to \code{label.x} and \code{label.y}, pass
+#'   also explicitly \code{geom = "text"}, \code{geom = "label"} or some other
+#'   geometry that use the \emph{x} and/or \emph{y} aesthetics. A vector with
+#'   the same length as the number of panels in the figure can be used if
+#'   needed.
 #'
 #' @section Computed variables: Data frame with one or more rows, one for each
 #'   group of observations for which counts are counted in \code{data}. \describe{
-#'   \item{npcx}{x value of label position in npc units}
-#'   \item{npcy}{y value of label position in npc units}
-#'   \item{count}{number of  observations}
-#'   \item{x}{x value of label position in data units}
-#'   \item{y}{y value of label position in data units}}.
+#'   \item{x,npcx}{x value of label position in data- or npc units, respectively}
+#'   \item{y,npcy}{y value of label position in data- or npc units, respectively}
+#'   \item{count}{number of  observations as an integer}}
 #'
 #'   As shown in one example below \code{\link[gginnards]{geom_debug}} can be
 #'   used to print the computed values returned by any statistic. The output
@@ -54,7 +56,7 @@
 #'   example. \code{x} and \code{y} are included in the output only if mapped.
 #'
 #' @return A plot layer instance. Using as output \code{data} the counts of
-#'   observations per plot quadrant.
+#'   observations in each plot panel or per group in each plot panel.
 #'
 #' @family Functions for quadrant and volcano plots
 #'
@@ -66,11 +68,24 @@
 #' set.seed(67821)
 #' x <- 1:100
 #' y <- rnorm(length(x), mean = 10)
-#' my.data <- data.frame(x, y)
+#' group <- factor(rep(c("A", "B"), times = 50))
+#' my.data <- data.frame(x, y, group)
 #'
 #' ggplot(my.data, aes(x, y)) +
 #'   geom_point() +
 #'   stat_panel_counts()
+#'
+#' ggplot(my.data, aes(x, y, colour = group)) +
+#'   geom_point() +
+#'   stat_panel_counts()
+#'
+#' ggplot(my.data, aes(x, y, colour = group)) +
+#'   geom_point() +
+#'   stat_group_counts()
+#'
+#' ggplot(my.data, aes(x, y, colour = group)) +
+#'   geom_point() +
+#'   stat_group_counts(label.x = "left", hstep = 0.06, vstep = 0)
 #'
 #' # We use geom_debug() to see the computed values
 #'
@@ -81,6 +96,11 @@
 #'   ggplot(my.data, aes(x, y)) +
 #'     geom_point() +
 #'     stat_panel_counts(geom = "debug")
+#'
+#'   ggplot(my.data, aes(x, y, colour = group)) +
+#'     geom_point() +
+#'     stat_group_counts(geom = "debug")
+#'
 #' }
 #'
 #' ggplot(my.data, aes(x, y)) +
@@ -92,6 +112,10 @@
 #'   stat_panel_counts(label.x = "left") +
 #'   stat_density()
 #'
+#' ggplot(my.data, aes(y, colour = group)) +
+#'   stat_group_counts(label.y = "top") +
+#'   stat_density(aes(fill = group))
+#'
 stat_panel_counts <- function(mapping = NULL,
                               data = NULL,
                               geom = "text_npc",
@@ -100,7 +124,8 @@ stat_panel_counts <- function(mapping = NULL,
                               label.y = "top",
                               na.rm = FALSE,
                               show.legend = FALSE,
-                              inherit.aes = TRUE, ...) {
+                              inherit.aes = TRUE,
+                              ...) {
 
   stopifnot(is.null(label.x) || is.numeric(label.x) || is.character(label.x))
   stopifnot(is.null(label.y) || is.numeric(label.y) || is.character(label.y))
@@ -116,6 +141,7 @@ stat_panel_counts <- function(mapping = NULL,
     params = list(na.rm = na.rm,
                   label.x = label.x,
                   label.y = label.y,
+                  npc.used = grepl("_npc", geom),
                   ...)
   )
 }
@@ -128,19 +154,67 @@ stat_panel_counts <- function(mapping = NULL,
 compute_panel_counts_fun <- function(data,
                                      scales,
                                      label.x,
-                                     label.y) {
+                                     label.y,
+                                     npc.used) {
 
   force(data)
 
   # total count
-  z <- tibble::tibble(count = nrow(data),
-                      npcx = label.x,
-                      npcy = label.y)
-  if ("x" %in% colnames(data)) {
-    z$x <- max(data$x) - 0.95 * diff(range(data$x))
+  z <- tibble::tibble(count = nrow(data))
+  # label position
+  if (is.character(label.x)) {
+    if (npc.used) {
+      margin.npc <- 0.05
+    } else {
+      # margin set by scale
+      margin.npc <- 0
+    }
+    label.x <- compute_npcx(x = label.x)
+    if (!npc.used) {
+      if ("x" %in% colnames(data)) {
+        x.expanse <- abs(diff(range(data$x)))
+        x.min <- min(data$x)
+        label.x <- label.x * x.expanse + x.min
+      } else {
+        if (data$PANEL[1] == 1L) { # show only once
+          message("No 'x' mapping; 'label.x' requires a numeric argument in data units")
+        }
+        label.x <- NA_real_
+      }
+    }
   }
-  if ("y" %in% colnames(data)) {
-    z$y <- max(data$y) - 0.95 * diff(range(data$y))
+  if (is.character(label.y)) {
+    if (npc.used) {
+      margin.npc <- 0.05
+    } else {
+      # margin set by scale
+      margin.npc <- 0
+    }
+    label.y <- compute_npcy(y = label.y)
+    if (!npc.used) {
+      if ("y" %in% colnames(data)) {
+        y.expanse <- abs(diff(range(data$y)))
+        y.min <- min(data$y)
+        label.y <- label.y * y.expanse + y.min
+      } else {
+        if (data$PANEL[1] == 1L) { # show only once
+          message("No 'y' mapping; 'label.y' requires a numeric argument in data units")
+        }
+        label.y <- NA_real_
+      }
+    }
+  }
+
+  if (npc.used) {
+    z$npcx <- label.x
+    z$x <- NA_real_
+    z$npcy <- label.y
+    z$y <- NA_real_
+  } else {
+    z$x <- label.x
+    z$npcx <- NA_real_
+    z$y <- label.y
+    z$npcy <- NA_real_
   }
   z
 }
@@ -152,6 +226,171 @@ compute_panel_counts_fun <- function(data,
 StatPanelCounts <-
   ggplot2::ggproto("StatPanelCounts", ggplot2::Stat,
                    compute_panel = compute_panel_counts_fun,
+                   default_aes =
+                     ggplot2::aes(npcx = ggplot2::after_stat(npcx),
+                                  npcy = ggplot2::after_stat(npcy),
+                                  label = sprintf("n=%i", ggplot2::after_stat(count)),
+                                  hjust = "inward",
+                                  vjust = "inward"),
+                   required_aes = c("x|y")
+  )
+
+
+#' @rdname stat_panel_counts
+#'
+#' @param hstep,vstep numeric in npc units, the horizontal and vertical step
+#'   used between labels for different groups.
+#'
+#' @export
+#'
+stat_group_counts <- function(mapping = NULL,
+                              data = NULL,
+                              geom = "text_npc",
+                              position = "identity",
+                              label.x = "right",
+                              label.y = "top",
+                              hstep = 0,
+                              vstep = NULL,
+                              na.rm = FALSE,
+                              show.legend = FALSE,
+                              inherit.aes = TRUE, ...) {
+
+  stopifnot(is.null(label.x) || is.numeric(label.x) || is.character(label.x))
+  stopifnot(is.null(label.y) || is.numeric(label.y) || is.character(label.y))
+
+  ggplot2::layer(
+    stat = StatGroupCounts,
+    data = data,
+    mapping = mapping,
+    geom = geom,
+    position = position,
+    show.legend = show.legend,
+    inherit.aes = inherit.aes,
+    params = list(na.rm = na.rm,
+                  label.x = label.x,
+                  label.y = label.y,
+                  hstep = hstep,
+                  vstep = ifelse(is.null(vstep),
+                                 ifelse(grepl("label", geom),
+                                        0.10,
+                                        0.05),
+                                 vstep),
+                  npc.used = grepl("_npc", geom),
+                  ...)
+  )
+}
+
+#' @rdname ggpp-ggproto
+#'
+#' @format NULL
+#' @usage NULL
+#'
+compute_group_counts_fun <- function(data,
+                                     params,
+                                     scales,
+                                     label.x,
+                                     label.y,
+                                     vstep,
+                                     hstep,
+                                     npc.used) {
+
+  force(data)
+
+  if (exists("grp.label", data)) {
+    if (length(unique(data[["grp.label"]])) > 1L) {
+      warning("Non-unique value in 'data$grp.label' using group index ", data[["group"]][1], " as label.")
+      grp.label <- as.character(data[["group"]][1])
+    } else {
+      grp.label <- data[["grp.label"]][1]
+    }
+  } else {
+    # if nothing mapped to grp.label we use group index as label
+    grp.label <- as.character(data[["group"]][1])
+  }
+
+  # Build group labels
+  group.idx <- abs(data$group[1])
+  if (length(label.x) >= group.idx) {
+    label.x <- label.x[group.idx]
+  } else if (length(label.x) > 0) {
+    label.x <- label.x[1]
+  }
+  if (length(label.y) >= group.idx) {
+    label.y <- label.y[group.idx]
+  } else if (length(label.y) > 0) {
+    label.y <- label.y[1]
+  }
+
+  # Compute number of observations
+  z <- tibble::tibble(count = nrow(data))
+
+  # Compute label positions
+  if (is.character(label.x)) {
+    if (npc.used) {
+      margin.npc <- 0.05
+    } else {
+      # margin set by scale
+      margin.npc <- 0
+    }
+    label.x <- compute_npcx(x = label.x, group = group.idx, h.step = hstep,
+                            margin.npc = margin.npc)
+    if (!npc.used) {
+      if ("x" %in% colnames(data)) {
+        x.expanse <- abs(diff(range(data$x)))
+        x.min <- min(data$x)
+        label.x <- label.x * x.expanse + x.min
+      } else {
+        if (data$PANEL[1] == 1L && group.idx == 1L) { # show only once
+          message("No 'x' mapping; 'label.x' requires a numeric argument in data units")
+        }
+        label.x <- NA_real_
+      }
+    }
+  }
+  if (is.character(label.y)) {
+    if (npc.used) {
+      margin.npc <- 0.05
+    } else {
+      # margin set by scale
+      margin.npc <- 0
+    }
+    label.y <- compute_npcy(y = label.y, group = group.idx, v.step = vstep,
+                            margin.npc = margin.npc)
+    if (!npc.used) {
+      if ("y" %in% colnames(data)) {
+        y.expanse <- abs(diff(range(data$y)))
+        y.min <- min(data$y)
+        label.y <- label.y * y.expanse + y.min
+      } else {
+        if (data$PANEL[1] == 1L && group.idx == 1L) { # show only once
+          message("No 'y' mapping; 'label.y' requires a numeric argument in data units")
+        }
+        label.y <- NA_real_
+      }
+    }
+  }
+
+  if (npc.used) {
+    z$npcx <- label.x
+    z$x <- NA_real_
+    z$npcy <- label.y
+    z$y <- NA_real_
+  } else {
+    z$x <- label.x
+    z$npcx <- NA_real_
+    z$y <- label.y
+    z$npcy <- NA_real_
+  }
+  z
+}
+
+#' @rdname ggpp-ggproto
+#' @format NULL
+#' @usage NULL
+#' @export
+StatGroupCounts <-
+  ggplot2::ggproto("StatGroupCounts", ggplot2::Stat,
+                   compute_group = compute_group_counts_fun,
                    default_aes =
                      ggplot2::aes(npcx = ggplot2::after_stat(npcx),
                                   npcy = ggplot2::after_stat(npcy),